Have fixed gd and sgd for ridge
The file gradient_descent_ols_and_ridge holds the code for gradient descent. I added a parameter to gd and sgd that allowed for changes to the regression model used. If we wanted to add a third rm in addition to ols and ridge, this would be an easy fix:)