Small changes
I  did an alternative preprocessing, because Im unsure if the original worked? I also checked to see what would happen if I removed all the rows with NaN in them. This removed approx half our data. I do not think we should do this. Instead we should work around NaN, because we have limited data. This can be a nice part of the discussion, where we say that we didnt want to lose to much data